+++ src/main/java/com/umulam/fleen/health/repository/jpa/BankAccountJpaRepository.java
@@ -7,15 +7,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface BankAccountJpaRepository extends JpaRepository<MemberBankAccount, Integer> {
+public interface BankAccountJpaRepository extends JpaRepository<MemberBankAccount, Long> {
 
   boolean existsByAccountNumber(String accountNumber);
   Optional<MemberBankAccount> findByAccountNumberAndMember(String accountNumber, Member member);
 
-  Optional<MemberBankAccount> findByIdAndMember(Integer id, Member member);
+  Optional<MemberBankAccount> findByIdAndMember(Long id, Member member);
 
   List<MemberBankAccount> findAllByMember(Member member);
 
-  boolean existsById(Integer id);
+  boolean existsById(Long id);
 
 }
