+++ src/main/java/com/umulam/fleen/health/repository/jpa/HealthSessionProfessionalJpaRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface HealthSessionProfessionalJpaRepository extends JpaRepository<Professional, Integer> {
+public interface HealthSessionProfessionalJpaRepository extends JpaRepository<Professional, Long> {
 
   @Query(
     value = "SELECT p FROM Professional p WHERE p.member.firstName LIKE CONCAT('%',INITCAP(:firstName),'%') " +
