+++ src/main/java/com/umulam/fleen/health/repository/jpa/MemberStatusJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberStatusJpaRepository extends JpaRepository<MemberStatus, Integer> {
+public interface MemberStatusJpaRepository extends JpaRepository<MemberStatus, Long> {
 
   Optional<MemberStatus> findByCode(String code);
 }
