+++ src/main/java/com/umulam/fleen/health/repository/jpa/transaction/WithdrawalTransactionJpaRepository.java
@@ -11,17 +11,17 @@ import org.springframework.web.bind.annotation.PathVariable;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface WithdrawalTransactionJpaRepository extends JpaRepository<WithdrawalTransaction, Integer> {
+public interface WithdrawalTransactionJpaRepository extends JpaRepository<WithdrawalTransaction, Long> {
 
   @Query(value = "SELECT wt FROM WithdrawalTransaction wt WHERE wt.reference = :reference")
   Optional<WithdrawalTransaction> findByReference(@Param("reference") String reference);
 
   @Query(value = "SELECT wt FROM WithdrawalTransaction wt WHERE wt.id = :transactionId AND wt.recipient.id = :memberId")
-  Optional<WithdrawalTransaction> findByUserAndId(@PathVariable("transactionId") Integer transactionId, @Param("memberId") Integer memberId);
+  Optional<WithdrawalTransaction> findByUserAndId(@PathVariable("transactionId") Long transactionId, @Param("memberId") Long memberId);
 
   @Query(value = "SELECT wt FROM WithdrawalTransaction wt WHERE wt.recipient.id = :memberId")
-  Page<WithdrawalTransaction> findAllByPayer(@Param("memberId") Integer memberId, Pageable pageable);
+  Page<WithdrawalTransaction> findAllByPayer(@Param("memberId") Long memberId, Pageable pageable);
 
   @Query(value = "SELECT wt FROM WithdrawalTransaction wt WHERE wt.recipient.id = :memberId AND wt.createdOn BETWEEN :startDate AND :endDate")
-  Page<WithdrawalTransaction> findByDateBetween(@Param("memberId") Integer memberId, @Param("startDate") LocalDateTime startDate, @Param("endDate") LocalDateTime endDate, Pageable pageable);
+  Page<WithdrawalTransaction> findByDateBetween(@Param("memberId") Long memberId, @Param("startDate") LocalDateTime startDate, @Param("endDate") LocalDateTime endDate, Pageable pageable);
 }
